trapèze presque pret

+++ arduino/asserv2/driver/rampe.h
@@ -21,33 +21,33 @@ class Rampe
 {
 	public:
 		Rampe();
-		void compute(long int actue, long int goal, long int speed, long int accel, long int decel);
+		void compute(double actue, double goal, double speed, double accel, double decel);
 		void compute_next_goal(long dt);
-		long get_goal();
+		double get_goal();
 		PHASE get_phase();
 
 	private:
 		int _sens;
 		/* U position à atteindre */
-		long int _goal;
-		/* U/µs vitesse à atteindre */
-		long int _speed;
-		/* U/µs² accélération voulue */
-		long int _acc;
-		/* U/µs² décélération voulue */
-		long int _dec;
+		double _goal;
+		/* U/ms vitesse à atteindre */
+		double _speed;
+		/* U/ms² accélération voulue */
+		double _acc;
+		/* U/ms² décélération voulue */
+		double _dec;
 		/* position actuelle dans la rampe */
-		long int _pos_actue;
+		double _pos_actue;
 		/* vitesse actuelle dans la rampe */
-		long int _speed_actue;
+		double _speed_actue;
 		/* accélération actuelle dans la rampe */
-		long int _acc_actue;
-		/* µs temps actuel */
-		long int _t;
-		/* µs temps de chaque étape */
-		long int _t01, _t12, _t23;
+		double _acc_actue;
+		/* ms temps actuel */
+		double _t;
+		/* ms temps de chaque étape */
+		double _t01, _t12, _t23;
 		/* U position à chaque changement détape */
-		long int _pos0, _pos1, _pos2, _pos3;
+		double _pos0, _pos1, _pos2, _pos3;
 		/* phase actuelle */
 		PHASE _phase;
 };
